WH-289: cc2538: Don't traverse boards subdirectory.
It seems putting the board-specific files in a separate .a file is
asking for trouble linker-wise.

automake will generate link requests roughly of the form:

	$(CC) -o $(output) $(objects) $(ldflags) $(libs)

and uses libtool to do it.

Trouble is as can be seen above, it separates linker flags from the
libraries, and will not let you mix the two.  So doing
`-Wl,--start-group lib1.a lib2.a lib3.a -Wl,--end-group` won't work.

I tried `--whole-file` at the start, and the linker blew up!

Ohh, and libtool is too clever for its own good, so specifying `lib1.a
lib2.a lib1.a` to get around a circular dependency, it'll assume the
second `lib1.a` was a mistake and omit it.

So, after that ramble, the solution: we bundle up the boards' files in
the platform .a file.  Less clean for sure, but it avoids this whole
dependency debarcle.