WH-289: cc2538 aes: Hash message length is in *bits*
I wondered why the length was given as a 64-bit quantity when the DMA
controller was effectively limited to a 32-bit length.

Doubly confused when you consider this is a Cortex M3 with mere
kilobytes of storage.  This is not a chip that will be streaming
NetFlicks in UHD.

Turns out, the message length is given in *bits*, not bytes.  So the
second length register is to hold the 3 bits that spill over the 32-bit
boundary.  If you have some obscenely (for this MCU) blob of data to
hash (more than 512MB), this lets you specify those upper 3 bits… and
theoretically more.